requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
cffi==1.11.5
cryptography==2.3
decorator==4.3.0
dj-database-url==0.5.0
Django==2.0.7
django-ckeditor==5.6.1
django-js-asset==1.1.0
idna==2.7
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.1
lxml==4.2.3
markdown2==2.3.5
mysqlclient==1.3.13
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.5
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
uWSGI==2.0.17.1
djangorestframework==3.8.2
django-cors-headers==2.4.0
django-filter==2.0.0
